Switched back to a simpler (clearer, easier-to-read) version of the triangular face computations for H(curl) pyramid.
I have now verified that we have agreement with ESEAS up to p=9, with the exception of some triangular-face dofs when these are evaluated at the apex.  Mathematically, these are 0, but ESEAS has values of order 1e-6 or so: a roundoff error of size 1e-18 gets magnified due to the way that the singularity of derivatives at the apex is handled numerically.  We have exact zeros thanks to some logic I added to the polynomial computations.