Ifpack2: added DRAFT support for custom overlap definitions using user-specified GIDs.
I think I don't yet have the interface for this quite right (the user has to supply GIDs as well as a local partitioning, whereas if the partitions are specified, the GIDs should be implied).

Also, it does not yet work!  Although my Camellia tests do solve the problem, they don't do so with the expected iteration counts; iteration counts are higher on 2 MPI processes than on 1.  By design, they should be identical, up to roundoff.

Signed-off-by: Nate Roberts <nvrober@sandia.gov>